+++ src/pyicloud_ipd/password_provider.py
@@ -1,8 +1,0 @@
-from enum import Enum
-
-
-class PasswordProvider(Enum):
-    PARAMETER = "parameter"
-    CONSOLE = "console"
-    KEYRING = "keyring"
-