+++ tests/test_authentication.py
@@ -5,12 +5,12 @@ import pytest
 from click.testing import CliRunner
 from icloudpd.logger import setup_logger
 import pyicloud_ipd
-from icloudpd.base import lp_filename_concatinator, main
+from icloudpd.base import dummy_password_writter, lp_filename_concatinator, main
 from icloudpd.authentication import authenticator, TwoStepAuthRequiredError
 import inspect
 
 from pyicloud_ipd.raw_policy import RawTreatmentPolicy
-from pyicloud_ipd.utils import identity
+from pyicloud_ipd.utils import constant, identity
 from tests.helpers import path_from_project_root, recreate_path
 
 vcr = VCR(decode_compressed_response=True)
@@ -35,7 +35,7 @@ class AuthenticationTestCase(TestCase):
             with self.assertRaises(
                 pyicloud_ipd.exceptions.PyiCloudFailedLoginException
             ) as context:
-                authenticator(setup_logger(), "com", identity, lp_filename_concatinator, RawTreatmentPolicy.AS_IS, {})(
+                authenticator(setup_logger(), "com", identity, lp_filename_concatinator, RawTreatmentPolicy.AS_IS, {"test": (constant("dummy"), dummy_password_writter)})(
                     "bad_username",
                     cookie_dir,
                     False,
@@ -58,7 +58,7 @@ class AuthenticationTestCase(TestCase):
                 # delete ./tests/vcr_cassettes/auth_requires_2sa.yml,
                 # put your actual credentials in here, run the test,
                 # and then replace with dummy credentials.
-                authenticator(setup_logger(), "com", identity, lp_filename_concatinator, RawTreatmentPolicy.AS_IS, {})(
+                authenticator(setup_logger(), "com", identity, lp_filename_concatinator, RawTreatmentPolicy.AS_IS, {"test": (constant("dummy"), dummy_password_writter)})(
                     "jdoe@gmail.com",
                     cookie_dir,
                     True,
@@ -83,7 +83,7 @@ class AuthenticationTestCase(TestCase):
                 # delete ./tests/vcr_cassettes/auth_requires_2fa.yml,
                 # put your actual credentials in here, run the test,
                 # and then replace with dummy credentials.
-                authenticator(setup_logger(), "com", identity, lp_filename_concatinator, RawTreatmentPolicy.AS_IS, {})(
+                authenticator(setup_logger(), "com", identity, lp_filename_concatinator, RawTreatmentPolicy.AS_IS, {"test": (constant("dummy"), dummy_password_writter)})(
                     "jdoe@gmail.com",
                     cookie_dir,
                     True,